requirements.txt
# Core dependencies
asyncio==3.4.3
cryptography==41.0.0
pyyaml==6.0
colorama==0.4.6
click==8.1.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0

# Development
black==23.0.0
flake8==6.0.0
mypy==1.0.0

# API & Frontend integration
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9